Create data_prep
In this project, a connection to an SQL database was established, and data from a dedicated database was extracted. First, the data was checked for any errors or missing values, and repairs were made where possible.

Next, four dataframes with analysis at the country-level granularity were prepared and merged into a single file, enabling further analysis and regression investigations. The same approach was applied to gather data at the club level.

Finally, the tables were exported using the to_pickle method to allow for easy use of the dataframes in another file dedicated to analysis and data visualization.